+++ client/components/Application/Dialog.tsx
@@ -1,7 +1,6 @@
 import * as store from '@client/store'
 import AddEmptyFolderDialog from './Dialogs/AddEmptyFolder'
 import CloseWithUnsavedChangesDialog from './Dialogs/CloseWithUnsavedChanges'
-import ConfigDialog from './Dialogs/Config'
 import CopyFileDialog from './Dialogs/CopyFile'
 import CopyFolderDialog from './Dialogs/CopyFolder'
 import DeleteFilesFoldersDialog from './Dialogs/DeleteFilesFolders'
@@ -27,8 +26,6 @@ export default function Dialog() {
 const DIALOGS = {
   addEmptyFolder: AddEmptyFolderDialog,
   closeWithUnsavedChanges: CloseWithUnsavedChangesDialog,
-  config: ConfigDialog,
-  configProject: ConfigDialog,
   copyFile: CopyFileDialog,
   copyFolder: CopyFolderDialog,
   deleteFilesFolders: DeleteFilesFoldersDialog,
