+++ client/store/actions/index.ts
@@ -1,12 +1,10 @@
 export * from './app'
-export * from './config'
 export * from './dialog'
 export * from './error'
 export * from './event'
 export * from './file'
 export * from './folder'
 export * from './panel'
-export * from './project'
 export * from './refs'
 export * from './resource'
 export * from './table'
