+++ client/store/state.ts
@@ -64,11 +64,6 @@ export type IState = {
   indexing?: boolean
 
   /**
-   * Application config object
-   **/
-  config?: types.IConfig
-
-  /**
    * Currently active event for example, file creation or deletion (to show animation)
    **/
   event?: types.IEvent
@@ -179,8 +174,6 @@ export type ITextState = {
 
 export type IDialog =
   | 'addEmptyFolder'
-  | 'config'
-  | 'configProject'
   | 'copyFile'
   | 'copyFolder'
   | 'create'
