--- server/endpoints/config/__init__.py
@@ -1,2 +1,0 @@
-# Register modules
-from . import read, write
