--- server/endpoints/config/read.py
@@ -1,28 +1,0 @@
-from __future__ import annotations
-
-from fastapi import Request
-from pydantic import BaseModel
-
-from ... import models
-from ...project import Project
-from ...router import router
-
-
-class Props(BaseModel, extra="forbid"):
-    pass
-
-
-class Result(BaseModel, extra="forbid"):
-    config: models.Config
-
-
-@router.post("/config/read")
-def endpoint(request: Request, props: Props) -> Result:
-    return action(request.app.get_project(), props)
-
-
-def action(project: Project, props: Props) -> Result:
-    cf = project.config
-
-    config = cf.read()
-    return Result(config=config)
